requirements.txt
google-adk
a2a-sdk
# Add any other dependencies your agent needs